# api/openapi.yaml
openapi: 3.1.0
info:
  title: llm-proxy web API
  version: 0.1.0
  description: |
    The browser-facing API of llm-proxy: the user cabinet and the admin panel.
    This is NOT the proxied LLM API (`/v1/...`), which the embedded upstream serves
    and which authenticates with API tokens, not sessions.

    Conventions:
    - Authentication is a session cookie set by `POST /api/auth/login` or the OIDC
      callback. It is `HttpOnly`, so no endpoint ever returns a session id.
    - A restricted session (temporary password not yet changed) may call only
      `GET /api/me`, `POST /api/auth/password` and `POST /api/auth/logout`; every
      other authenticated endpoint answers 403 with code `password_change_required`.
    - An authenticated endpoint called without a valid session answers 401 with code
      `unauthenticated`. The only other 401 code is `invalid_credentials`, and it is
      returned only by `POST /api/auth/login` and `POST /api/auth/password`, where it
      rejects what was typed and leaves the session alone: a client must not treat it
      as a signed-out session.
    - Every error the web API writes, including those from its middleware (session,
      restriction, content type), is a JSON `Error` - never plain text.
    - Codes any endpoint may answer besides those it declares: `unsupported_media_type`
      (415, a mutating request without `Content-Type: application/json`),
...
      expects, with `field` when one field is at fault).
...
servers:
  - url: /
...
paths:
...
  /api/me/tokens:
    get:
      tags: [cabinet]
      operationId: listMyTokens
      responses:
        "200":
          description: The caller's API tokens, newest first, revoked ones included.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Token" }
    post:
      tags: [cabinet]
      operationId: issueMyToken
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueTokenRequest" }
      responses:
        "201":
          description: Issued. `secret` is shown here and never again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssuedToken" }
        "422":
          description: "`invalid_input` with `field: label` — empty, longer than 64 characters, or with control characters."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`token_limit` — the account already holds 50 tokens that are not revoked; revoke one first."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/me/tokens/{tokenId}:
    delete:
      tags: [cabinet]
      operationId: revokeMyToken
      parameters:
        - $ref: "#/components/parameters/TokenId"
      responses:
        "204": { description: Revoked. Takes effect on the next proxied request. }
        "404":
          description: "`not_found` — also for a token owned by someone else."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/me/usage:
    get:
      tags: [cabinet]
      operationId: getMyUsage
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Aggregated consumption for the caller.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usage" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: llmproxy_session
  parameters:
...
    TokenId: { name: tokenId, in: path, required: true, schema: { type: string, format: uuid } }
...
    From: { name: from, in: query, schema: { type: string, format: date-time }, description: Defaults to 7 days before `to`. }
    To: { name: to, in: query, schema: { type: string, format: date-time }, description: Defaults to now. }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code. The client translates this, not `message`.
          examples: [invalid_credentials]
        message: { type: string, description: English diagnostic for logs. Not for display. }
        field: { type: string, description: "The request field the error is about, when there is one." }
...
    Token:
      type: object
      required: [id, label, prefix, createdAt]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        prefix: { type: string, description: "The first characters of the secret, to tell keys apart." }
        createdAt: { type: string, format: date-time }
        lastUsedAt:
          type: [string, "null"]
          format: date-time
          description: Null if the token was never used.
        revokedAt: { type: [string, "null"], format: date-time }
    IssueTokenRequest:
      type: object
      required: [label]
      properties:
        label: { type: string, minLength: 1, maxLength: 64, description: "Printable text; control characters are refused." }
    IssuedToken:
      type: object
      required: [token, secret]
      properties:
        token: { $ref: "#/components/schemas/Token" }
        secret: { type: string, description: The full API token. Shown once. }
    Usage:
      type: object
      required: [from, to, bucket, totals, points]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        bucket: { type: string, enum: [hour, day] }
        totals:
          type: object
          required: [requests, tokensTotal]
          properties:
            requests: { type: integer, description: "Served requests. An attempt the gateway retried on another account is not counted twice; failed attempts are not counted." }
            tokensTotal: { type: integer, description: "Every token spent, including tokens of failed or retried attempts." }
        points:
          type: array
          items:
            type: object
            required: [at, model, requests, tokensTotal]
            properties:
              at: { type: string, format: date-time, description: Start of the bucket. }
              model: { type: string }
              requests: { type: integer, description: "Served requests in this bucket, as in `totals.requests`." }
              tokensTotal: { type: integer }
...
security:
  - session: []